finished navbar basics
