+++ frontend/src/App.js
@@ -27,7 +27,7 @@ function App() {
 				<Route path="/signup">
 					<SignupFormPage />
 				</Route>
-				<Route path="/desktop">
+				<Route path={["/desktop", "/notes", "/notebooks", "/trash"]}>
 					<DesktopPage />
 				</Route>
 			</Switch>
