+++ frontend/src/components/DesktopPage/desktop.js
@@ -1,0 +1,9 @@
+const Desktop = () => {
+  return (
+    <main>
+      <h1>Desktop</h1>
+    </main>
+  )
+}
+
+export default Desktop;
