+++ frontend/src/components/DesktopPage/notebooks.js
@@ -1,0 +1,9 @@
+const NotebookPage = ({notebooks}) => {
+  return(
+    <main>
+      <h1>Notebook Page</h1>
+    </main>
+  )
+};
+
+export default NotebookPage;
