+++ frontend/src/components/DesktopPage/notes.js
@@ -1,0 +1,9 @@
+const NotesPage = ({notes}) => {
+  return(
+    <main>
+      <h1>Notes</h1>
+    </main>
+  )
+};
+
+export default NotesPage;
