+++ frontend/src/components/DesktopPage/trash.js
@@ -1,0 +1,9 @@
+const TrashPage = ({trash}) => {
+  return(
+    <main>
+      <h1>Trash Page</h1>
+    </main>
+  )
+}
+
+export default TrashPage;
