+++ frontend/src/components/Navigation/trash.js
@@ -1,0 +1,14 @@
+const Trash = () => {
+
+  return (
+    <div className="nav-div">
+      <div className="nav-div-left">
+        <div className="nav-caret"></div>
+        <i className="fa-solid fa-trash-can"></i>
+        <div>Trash</div>
+      </div>
+		</div>
+  )
+};
+
+export default Trash;
