+++ frontend/src/store/index.js
@@ -4,13 +4,15 @@ import thunk from "redux-thunk";
 import sessionReducer from "./session";
 import notebookReducer from "./notebooks";
 import noteReducer from "./notes";
-import tagReducer from "./tags"
+import tagReducer from "./tags";
+import trashReducer from "./trash";
 
 const rootReducer = combineReducers({
 	session: sessionReducer,
   notebooks: notebookReducer,
   notes: noteReducer,
   tags: tagReducer,
+  trash: trashReducer,
 });
 
 let enhancer;
