+++ frontend/src/store/session.js
@@ -1,8 +1,8 @@
 import { csrfFetch } from "./csrf";
 
 // Actions
-const SET_USER = "session/setUser";
-const REMOVE_USER = "session/removeUser";
+const SET_USER = "session/SET_USER";
+const REMOVE_USER = "session/REMOVE_USER";
 
 // Action Creators
 const setUser = (user) => {
